zookeeper boot script

+++ zookeeper/start-zk.sh
@@ -1,24 +1,37 @@
 #!/bin/sh
 hostname=`hostname`
+#hostname='zk1.svc'
+is_exist=0
 
-`etcdctl2 ls /zk/hosts/${hostname} > /dev/null 2>&1 `  || 
-confd -confdir=/bootstrap/zookeeper/confd/ -watch=true -onetime=false -backend etcd -node http://etcd.svc:2379 & \
-&& /usr/local/zookeeper/bin/zkServer.sh start-foreground \
-&& \
-exit 'exists hostname'
+`etcdctl2 ls /zk/hosts/${hostname} > /dev/null 2>&1 ` && is_exist=1
 
-count=`etcdctl2 ls /zk/hosts 2> /dev/null | wc -l`
-id=$(($count+1))
+if [ $is_exist -eq 1 ] ; 
+then
+  echo 'hostname already exists!!!!!!!!'
+  confd -confdir=/bootstrap/zookeeper/confd/ -watch=true -onetime=false -backend etcd -node http://etcd.svc:2379 & 
+  /bootstrap/zookeeper/reload
+  #/usr/local/zookeeper/bin/zkServer.sh start-foreground 
+else
+  
+  count=`etcdctl2 ls /zk/hosts 2> /dev/null | wc -l`
+  id=$(($count+1))
+  
+  echo $count
 
-etcdctl2 set /zk/hosts/${hostname} '{"id":"'${id}'","hostname":"'${hostname}'"}'
+  confd -confdir=/bootstrap/zookeeper/confd/ -watch=true -onetime=false -backend etcd -node http://etcd.svc:2379 &
+
+  etcdctl2 set /zk/hosts/${hostname} '{"id":"'${id}'","hostname":"'${hostname}'"}'
+  
+  rm -rf /tmp/zookeeper/ 2> /dev/null
+  mkdir -p /tmp/zookeeper/logs /tmp/zookeeper/zkdata
+  echo "${id}" > /tmp/zookeeper/zkdata/myid
+  
+  etcdctl2 set /trackdown/${hostname} "/zk/hosts/${hostname}"
+  
+fi
 
 
-
-confd -confdir=/bootstrap/zookeeper/confd/ -watch=true -onetime=false -backend etcd -node http://etcd.svc:2379 &
-rm -rf /tmp/zookeeper/ 2> /dev/null
-mkdir -p /tmp/zookeeper/logs /tmp/zookeeper/zkdata
-echo "${id}" > /tmp/zookeeper/zkdata/myid
-
-etcdctl2 set /trackdown/${hostname} "/zk/hosts/${hostname}"
-
-/usr/local/zookeeper/bin/zkServer.sh start-foreground
+while true
+do
+    sleep 86400
+done
